truncate machine-id on template creation

+++ clout_init_tpl.sh
@@ -32,12 +32,14 @@ fi
 mkdir -p /mnt/customize_image
 guestmount -a $IMAGE -i --rw /mnt/customize_image
 cd /mnt/customize_image/var/lib/systemd/
-dd if=/dev/urandom of=random-seed bs=512 count=4
-chmod 755 random-seed
+if [ ! -f random-seed ]; then
+  dd if=/dev/urandom of=random-seed bs=512 count=4
+  chmod 755 random-seed
+fi
 cd -
 guestunmount /mnt/customize_image/
 
-virt-customize -a $IMAGE --install qemu-guest-agent
+virt-customize -a $IMAGE --install qemu-guest-agent --truncate /etc/machine-id
 
 qm create $VM_ID --name $TEMPLATE_NAME --memory $MEMORY --net0 virtio,bridge=$NET_BRIDGE --cpu host --sockets 1 --cores $CPU
